requirements.txt
streamlit~=1.42.0
openai~=1.61.1
PyPDF2~=3.0.1
python-dotenv~=1.0.1